svg/geometry/parsing/rx-valid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="8000"><title>SVG Geometry Properties: parsing rx with valid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/geometry.html#RX" rel="help"/><h:meta content="rx supports the full grammar '&lt;length-percentage&gt; | auto'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_valid_value("rx", "auto");

test_valid_value("rx", "0", "0px");
test_valid_value("rx", "1px");
test_valid_value("rx", "calc(2em + 3ex)");
test_valid_value("rx", "4%");
test_valid_value("rx", "5vw");

  ]]></script></svg>
